% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USbreweriesdata.R
\name{get_brewery}
\alias{get_brewery}
\title{Breweries in U.S}
\usage{
get_brewery(key)
}
\arguments{
\item{key}{BreweryBP API key}
}
\description{
This function makes it easy to create a dataframe of different breweries
in the U.S and their location inforamtion using the BreweryBP sandbox API data.
}
\details{
The function asks the user to provide their BreweryBP API key, and then it will return a small dataframe of brewery location information.
}
\examples{
get_brewery(key = "BreweryDB_API")
}
\keyword{API}
\keyword{Brewery}
\keyword{base_url}
\keyword{df}
\keyword{key}
\keyword{key_preface}
\keyword{locations}
