% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USbreweriesbystate.R
\name{get_brewery_bystate}
\alias{get_brewery_bystate}
\title{Breweries by State}
\usage{
get_brewery_bystate(state, key)
}
\arguments{
\item{state}{A state in the U.S}

\item{key}{BreweryBP API key}
}
\description{
This function makes it easy to filter out breweries and their location information
by state using the BreweryBP sandbox API data..
}
\details{
The function asks the user to provide their BreweryBP API key
and the state they wish to search, and then it will return a small
dataframe of brewery location information by state.
}
\examples{
get_brewery_bystate(state = "", key = "BreweryDB_API")

}
\keyword{API}
\keyword{Brewery}
\keyword{base_url}
\keyword{df}
\keyword{key}
\keyword{key_preface}
\keyword{locations}
\keyword{state}
