% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USbreweryhomevalues.R
\name{get_breweryhomeinfo}
\alias{get_breweryhomeinfo}
\title{Brewery & Home Info}
\usage{
get_breweryhomeinfo(state)
}
\arguments{
\item{state}{A state in the U.S}
}
\description{
This function makes it easy to search home sales and rental
valuations by breweries in different states.
}
\details{
The function asks the user to provide a U.S state that they would
like to search for information on home sale and rental valuations
around breweries.
}
\examples{
get_breweryhomeinfo(state = "Colorado")

}
\keyword{Brewery}
\keyword{home}
\keyword{locations}
\keyword{rentals}
\keyword{sales}
\keyword{state}
