% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USbrewerymap.R
\name{plot_breweryhomemap}
\alias{plot_breweryhomemap}
\title{Map Breweries & Home Info}
\usage{
plot_breweryhomemap()
}
\description{
This function creates an interactive leaflet map of the different brewery locations,
which includes popups with informaiton on the breweries and home sale and rental
information around each brewery location.
}
\details{
The function does not have any parameters, it will plot the interactive map
when the function is called.
}
\examples{
plot_breweryhomemap()

}
\keyword{Brewery}
\keyword{home}
\keyword{locations}
\keyword{rentals}
\keyword{sales}
\keyword{state}
